chore: move maven install to a pkg
