+++ internal/librarian/java/install.go
@@ -21,12 +21,10 @@ import (
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strings"
 
 	"github.com/googleapis/librarian/internal/cache"
-	"github.com/googleapis/librarian/internal/command"
 	"github.com/googleapis/librarian/internal/config"
-	"github.com/googleapis/librarian/internal/filesystem"
+	"github.com/googleapis/librarian/internal/tool/maven"
 )
 
 const (
@@ -64,18 +62,7 @@ func Install(ctx context.Context, tools *config.Tools) error {
 	if err := os.MkdirAll(libDir, 0o755); err != nil {
 		return fmt.Errorf("failed to create lib directory %q: %w", libDir, err)
 	}
-	for _, mvnTool := range tools.Maven {
-		var err error
-		if mvnTool.LocalPath != "" {
-			err = installLocalMavenTool(ctx, mvnTool, binDir, libDir)
-		} else {
-			err = installExternalMavenTool(ctx, mvnTool, binDir, libDir)
-		}
-		if err != nil {
-			return fmt.Errorf("failed to install maven tool %s: %w", mvnTool.Name, err)
-		}
-	}
-	return nil
+	return maven.Install(ctx, tools.Maven, binDir, libDir)
 }
 
 // InstallDir returns the absolute path of the installation directory for Java tools.
@@ -85,155 +72,6 @@ func InstallDir() (string, error) {
 		return "", err
 	}
 	return filepath.Abs(filepath.Join(dir, toolsDir))
-}
-
-// installExternalMavenTool downloads a Maven-based external tool, copies its compiled artifact
-// (.jar or .exe) to the sibling lib folder, and creates an executable wrapper script
-// in the bin folder pointing directly to that library file.
-func installExternalMavenTool(ctx context.Context, mvnTool *config.MavenTool, binDir, libDir string) error {
-	artifact, ext := getM2ArtifactSpec(mvnTool)
-	if err := downloadM2Artifact(ctx, artifact, binDir); err != nil {
-		return err
-	}
-	artifactPath, err := resolveM2ArtifactPath(mvnTool, ext)
-	if err != nil {
-		return err
-	}
-	if _, err := os.Stat(artifactPath); err != nil {
-		return fmt.Errorf("downloaded artifact not found at %s: %w", artifactPath, err)
-	}
-	isExe := ext == "exe"
-	destPath, err := copyArtifactToLib(artifactPath, libDir, isExe)
-	if err != nil {
-		return err
-	}
-	return createBinWrapper(mvnTool.Name, destPath, binDir, isExe, mvnTool.MainClass)
-}
-
-// installLocalMavenTool compiles a local Maven project, parses its pom.xml metadata coordinates,
-// copies the built target artifact (.jar or .exe) to the sibling lib folder, and creates an executable
-// wrapper script in the bin folder.
-func installLocalMavenTool(ctx context.Context, mvnTool *config.MavenTool, binDir, libDir string) error {
-	absLocalPath, err := filepath.Abs(mvnTool.LocalPath)
-	if err != nil {
-		return fmt.Errorf("failed to resolve absolute local path for %s: %w", mvnTool.LocalPath, err)
-	}
-	if err := buildLocalMavenProject(ctx, mvnTool.LocalPath); err != nil {
-		return err
-	}
-	pomPath := filepath.Join(absLocalPath, "pom.xml")
-	proj, err := parsePOM(pomPath)
-	if err != nil {
-		return err
-	}
-	ext := mvnTool.Packaging
-	if ext == "" {
-		ext = "jar"
-	}
-	fileName := fmt.Sprintf("%s-%s.%s", proj.ArtifactID, proj.Version, ext)
-	artifactPath := filepath.Join(absLocalPath, "target", fileName)
-	if _, err := os.Stat(artifactPath); err != nil {
-		return fmt.Errorf("compiled artifact not found at %q: %w", artifactPath, err)
-	}
-	isExe := ext == "exe"
-	destPath, err := copyArtifactToLib(artifactPath, libDir, isExe)
-	if err != nil {
-		return err
-	}
-	return createBinWrapper(mvnTool.Name, destPath, binDir, isExe, mvnTool.MainClass)
-}
-
-// getM2ArtifactSpec constructs the Maven coordinate string and returns it along with the file extension.
-func getM2ArtifactSpec(mvnTool *config.MavenTool) (string, string) {
-	ext := mvnTool.Packaging
-	if ext == "" {
-		ext = "jar"
-	}
-	artifact := fmt.Sprintf("%s:%s:%s:%s", mvnTool.GroupID, mvnTool.ArtifactID, mvnTool.Version, ext)
-	if mvnTool.Classifier != "" {
-		artifact = fmt.Sprintf("%s:%s", artifact, mvnTool.Classifier)
-	}
-	return artifact, ext
-}
-
-// downloadM2Artifact executes mvn dependency:get to download the target artifact.
-func downloadM2Artifact(ctx context.Context, artifact, workDir string) error {
-	args := []string{
-		"dependency:get",
-		"-Dartifact=" + artifact,
-	}
-	if err := command.RunStreamingInDir(ctx, workDir, "mvn", args...); err != nil {
-		return fmt.Errorf("failed to download artifact %s: %w", artifact, err)
-	}
-	return nil
-}
-
-// resolveM2ArtifactPath returns the absolute path to the downloaded artifact in the local .m2 repository.
-func resolveM2ArtifactPath(mvnTool *config.MavenTool, ext string) (string, error) {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return "", fmt.Errorf("failed to get user home directory: %w", err)
-	}
-	m2Repo := filepath.Join(home, ".m2", "repository")
-	groupIDPath := strings.ReplaceAll(mvnTool.GroupID, ".", "/")
-	fileName := fmt.Sprintf("%s-%s", mvnTool.ArtifactID, mvnTool.Version)
-	if mvnTool.Classifier != "" {
-		fileName = fmt.Sprintf("%s-%s", fileName, mvnTool.Classifier)
-	}
-	fileName = fmt.Sprintf("%s.%s", fileName, ext)
-	return filepath.Join(m2Repo, groupIDPath, mvnTool.ArtifactID, mvnTool.Version, fileName), nil
-}
-
-// copyArtifactToLib copies the artifact file into the isolated sibling lib directory,
-// applying execution permission bits if needed.
-func copyArtifactToLib(srcPath, libDir string, makeExecutable bool) (string, error) {
-	fileName := filepath.Base(srcPath)
-	destPath := filepath.Join(libDir, fileName)
-	if err := filesystem.CopyFile(srcPath, destPath); err != nil {
-		return "", fmt.Errorf("failed to copy artifact to lib folder: %w", err)
-	}
-	if makeExecutable {
-		if err := os.Chmod(destPath, 0o755); err != nil {
-			return "", fmt.Errorf("failed to make copied exe executable: %w", err)
-		}
-	}
-	return destPath, nil
-}
-
-// createBinWrapper creates a shell wrapper script in the bin directory that forwards executions to the library file.
-func createBinWrapper(wrapperName, destPath, binDir string, isExecutable bool, mainClass string) error {
-	wrapperPath := filepath.Join(binDir, wrapperName)
-	var content string
-	switch {
-	case isExecutable:
-		content = fmt.Sprintf("#!/bin/sh\nexec %q \"$@\"\n", destPath)
-	case mainClass != "":
-		content = fmt.Sprintf("#!/bin/sh\nexec java -cp %q %q \"$@\"\n", destPath, mainClass)
-	default:
-		content = fmt.Sprintf("#!/bin/sh\nexec java -jar %q \"$@\"\n", destPath)
-	}
-	return os.WriteFile(wrapperPath, []byte(content), 0o755)
-}
-
-// buildLocalMavenProject builds the local Maven project at the target relative path under the monorepo root.
-func buildLocalMavenProject(ctx context.Context, localPath string) error {
-	args := []string{
-		"package",
-		"-B",
-		"-ntp",
-		"-T", "1.5C",
-		"-DskipTests",
-		"-Dcheckstyle.skip",
-		"-Dclirr.skip",
-		"-Denforcer.skip",
-		"-Dfmt.skip",
-		"-pl", localPath,
-		"--also-make",
-	}
-	if err := command.RunStreaming(ctx, "mvn", args...); err != nil {
-		return fmt.Errorf("failed to build local Maven project %q: %w", localPath, err)
-	}
-	return nil
 }
 
 // getBinDir returns the absolute path of the directory where Java tool wrapper scripts are stored.
